+++ lib/src/mock_store.dart
@@ -27,9 +27,11 @@ class MockStore<St> extends Store<St> {
     ModelObserver? modelObserver,
     ErrorObserver<St>? errorObserver,
     WrapReduce<St>? wrapReduce,
-    @Deprecated("Use `globalWrapError` instead. This will be removed.")
-    WrapError<St>? wrapError,
+    GlobalErrorObserver<St> Function()? globalErrorObserver,
+    //
+    @Deprecated("Use `globalErrorObserver` instead. This will be removed.")
     GlobalWrapError<St>? globalWrapError,
+    //
     bool? defaultDistinct,
     CompareBy? immutableCollectionEquality,
     int? maxErrorsQueued,
@@ -47,7 +49,7 @@ class MockStore<St> extends Store<St> {
           modelObserver: modelObserver,
           errorObserver: errorObserver,
           wrapReduce: wrapReduce,
-          wrapError: wrapError,
+          globalErrorObserver: globalErrorObserver,
           globalWrapError: globalWrapError,
           defaultDistinct: defaultDistinct,
           immutableCollectionEquality: immutableCollectionEquality,
@@ -110,8 +112,7 @@ class MockStore<St> extends Store<St> {
 
   @Deprecated("Use `dispatchAndWait` instead. This will be removed.")
   @override
-  Future<ActionStatus> dispatchAsync(ReduxAction<St> action,
-      {bool notify = true}) {
+  Future<ActionStatus> dispatchAsync(ReduxAction<St> action, {bool notify = true}) {
     return dispatchAndWait(action, notify: notify);
   }
 
@@ -139,8 +140,7 @@ class MockStore<St> extends Store<St> {
   /// - [dispatch] which dispatches both sync and async actions.
   /// - [dispatchSync] which dispatches sync actions, and throws if the action is async.
   @override
-  Future<ActionStatus> dispatchAndWait(ReduxAction<St> action,
-      {bool notify = true}) {
+  Future<ActionStatus> dispatchAndWait(ReduxAction<St> action, {bool notify = true}) {
     ReduxAction<St>? _action = _getMockedAction(action);
 
     return (_action == null) //